Added etcdutility
This tool was primarely used as a quick check, if any node does not
override the retry value in etcd. Because for every new node the retry
value was set to 600 by the original python script while the default value
was set to 20. Before changing the default value and removing all
overrides, this tool was created to check, if this would affect any node